# app/api/transactions/route.ts
import { withLedgerWriteAudit } from '../../../lib/ledgerContext';
import { NextRequest, NextResponse } from 'next/server';
import { requireAuth } from '../../../lib/apiHelpers';
import { queryAll, queryOne } from '../../../lib/db';
import { normalizeCurrency, convertToTwd, normalizeDate, resolveOverseasFee } from '../../../lib/accountHelpers';
import { todayInUserTz, isValidIsoDate } from '../../../lib/userTime';
import { computeTwdAmount } from '../../../lib/moneyDecimal';
import { insertIncomeExpenseTransaction } from '../../../lib/transactionWriteCore';
import { emitTransactionEvent } from '../../../lib/transactionWebhooks';

type TransactionType = 'income' | 'expense' | 'transfer_in' | 'transfer_out';
type SortField = 'date' | 'amount' | 'account' | 'category' | 'type';
type SortDir = 'ASC' | 'DESC';

interface TransactionRow {
  id: string;
  user_id: string;
  type: TransactionType | string;
  amount: number;
  currency: string | null;
  original_amount: number | null;
  fx_rate: string | number | null;
  fx_fee: number | null;
  twd_amount: number | null;
  date: string;
  category_id: string | null;
  account_id: string | null;
  transfer_to_account_id: string | null;
  note: string | null;
  exclude_from_stats: number | null;
  is_fx_fee: number | null;
  linked_id: string | null;
  source_recurring_id: string | null;
  source_recurring_name?: string | null;
  scheduled_date: string | null;
  ai_created: number | null;
  note_ai_modified: number | null;
  created_at: string | number | null;
  updated_at: string | number | null;
  [key: string]: unknown;
}

interface TransactionListItem extends TransactionRow {
  categoryId: string | null;
  accountId: string | null;
  toAccountId: string | null;
  currency: string;
  originalAmount: number;
  fxRate: string;
  fxFee: number;
  twdAmount: number;
  excludeFromStats: boolean;
  isFxFee: boolean;
  aiCreated: boolean;
  noteAiModified: boolean;
  repaymentSummaryId: string;
  linkedId: string;
  sourceRecurringId: string | null;
  sourceRecurringName: string | null;
  scheduledDate: string | null;
  createdAt: string | number | null;
  updatedAt: string | number | null;
  attachmentCount: number;
  firstAttachmentId: string | null;
}

interface TransactionListResponse {
  data: TransactionListItem[];
  items: TransactionListItem[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
  sort: string;
}

interface CreateTransactionRequest {
  type?: string;
  amount?: number | string;
  originalAmount?: number | string;
  currency?: string;
  fxRate?: number | string | null;
  fxFee?: number | string | null;
  date?: string | null;
  categoryId?: string | null;
  accountId?: string | null;
  note?: string | null;
  excludeFromStats?: boolean;
  /** 離線記帳的 idempotency key（007）；線上直接新增時省略。 */
  clientRef?: string | null;
}

/** 離線 idempotency key 形狀驗證：限 32 碼十六進位，避免使用者塞任意長字串。 */
const CLIENT_REF_REGEX = /^[a-f0-9]{32}$/;

const SORT_REGEX = /^(date|amount|account|category|type)_(asc|desc)$/;
const TRANSACTION_TYPES = new Set(['income', 'expense', 'transfer_in', 'transfer_out']);

function asRows<T>(rows: Array<Record<string, string | number | null>>): T[] {
  return rows as unknown as T[];
}

function asRow<T>(row: Record<string, string | number | null> | null): T | null {
  return row as unknown as T | null;
}

export async function GET(request: NextRequest) {
  const auth = await requireAuth(request);
  if (auth instanceof NextResponse) return auth;

  const { searchParams } = new URL(request.url);
  const dateFrom = searchParams.get('dateFrom') || '';
  const dateTo = searchParams.get('dateTo') || '';
  const type = searchParams.get('type') || '';
  const categoryId = searchParams.get('categoryId') || '';
  const accountId = searchParams.get('accountId') || '';
  const page = searchParams.get('page') || '1';
  const keyword = String(searchParams.get('keyword') || '').trim();
  const excludeTransferRaw = String(searchParams.get('excludeTransfer') || '').toLowerCase();
  const excludeTransfer = excludeTransferRaw === '1' || excludeTransferRaw === 'true';

  const limitRaw = parseInt(searchParams.get('limit') || '', 10);
  const pageSize = Number.isFinite(limitRaw) && limitRaw > 0 ? limitRaw : 20;
  if (pageSize > 500) {
    return NextResponse.json({ error: '每頁最多 500 筆', code: 'PageSizeOutOfRange' }, { status: 400 });
  }

  const sortStr = String(searchParams.get('sort') || 'date_desc').toLowerCase();
  const sortMatch = SORT_REGEX.exec(sortStr);
  if (searchParams.get('sort') && !sortMatch) {
    return NextResponse.json({ error: 'sort 參數格式無效', code: 'ValidationError', field: 'sort' }, { status: 400 });
  }
  const sortField = (sortMatch ? sortMatch[1] : 'date') as SortField;
  const sortDir: SortDir = sortMatch && sortMatch[2] === 'asc' ? 'ASC' : 'DESC';

  const needJoinAcc = sortField === 'account';
  const needJoinCat = sortField === 'category';
  let baseTable = 'transactions t';
  if (needJoinAcc) baseTable += ' LEFT JOIN accounts acc ON acc.id = t.account_id';
  if (needJoinCat) baseTable += ' LEFT JOIN categories cat ON cat.id = t.category_id';
  baseTable += ' LEFT JOIN recurring r ON r.id = t.source_recurring_id AND r.user_id = t.user_id';

  const today = todayInUserTz(auth.userTimezone);
  let where = 't.user_id = ?';
  const params: Array<string | number | null> = [auth.userId];

  if (dateFrom) { where += ' AND t.date >= ?'; params.push(dateFrom); }
  if (dateTo) { where += ' AND t.date <= ?'; params.push(dateTo); }
  if (type && type !== 'all') {
    if (type === 'transfer') {
      where += " AND (t.type = 'transfer_out' OR t.type = 'transfer_in')";
    } else if (type === 'future') {
      where += ' AND t.date > ?';
      params.push(today);
    } else {
      where += ' AND t.type = ?'; params.push(type);
    }
  }
  // 排除轉帳：轉入/轉出兩腳都不列出（與 type=transfer 併用時必然無結果，由前端互斥處理）。
  if (excludeTransfer) {
    where += " AND t.type NOT IN ('transfer_in', 'transfer_out')";
  }
  if (categoryId && categoryId !== 'all') {
    if (categoryId === '__uncategorized__') {
      where += " AND (t.category_id IS NULL OR t.category_id = '') AND t.type IN ('income', 'expense') AND COALESCE(t.exclude_from_stats, 0) = 0";
    } else {
    const requested = String(categoryId).split(',').map(s => s.trim()).filter(Boolean);
    if (requested.length > 0) {
      const placeholders = requested.map(() => '?').join(',');
      const childRows = asRows<{ id: string }>(queryAll(
        `SELECT id FROM categories WHERE user_id = ? AND parent_id IN (${placeholders})`,
        [auth.userId, ...requested]
      ));
      const expanded = new Set(requested);
      childRows.forEach(r => expanded.add(r.id));
      const ids = [...expanded];
      const idPh = ids.map(() => '?').join(',');
      where += ` AND t.category_id IN (${idPh})`;
      params.push(...ids);
    }
    }
  }
  if (accountId && accountId !== 'all') { where += ' AND t.account_id = ?'; params.push(accountId); }
  if (keyword) { where += ' AND LOWER(t.note) LIKE LOWER(?)'; params.push(`%${keyword}%`); }

  const countSql = `SELECT COUNT(*) as cnt FROM ${baseTable} WHERE ${where}`;
  const total = Number(asRow<{ cnt: number }>(queryOne(countSql, params))?.cnt) || 0;

  let orderClause = '';
  if (sortField === 'date') orderClause = `ORDER BY t.date ${sortDir}, t.created_at DESC`;
  else if (sortField === 'amount') orderClause = `ORDER BY t.amount ${sortDir}, t.date DESC`;
  else if (sortField === 'type') orderClause = `ORDER BY t.type ${sortDir}, t.date DESC`;
  else if (sortField === 'account') orderClause = `ORDER BY acc.name ${sortDir}, t.date DESC`;
  else if (sortField === 'category') orderClause = `ORDER BY cat.name ${sortDir}, t.date DESC`;

  const pageNum = parseInt(page, 10) || 1;
  const offset = (pageNum - 1) * pageSize;
  const selectCols = `t.*,
    COALESCE(NULLIF(r.note, ''), '（未命名配方）') AS source_recurring_name,
    (SELECT COUNT(*) FROM transaction_attachments ta WHERE ta.transaction_id = t.id AND ta.user_id = t.user_id) AS attachment_count,
    (SELECT ta.id FROM transaction_attachments ta WHERE ta.transaction_id = t.id AND ta.user_id = t.user_id ORDER BY ta.created_at ASC LIMIT 1) AS first_attachment_id`;
  const sql = `SELECT ${selectCols} FROM ${baseTable} WHERE ${where} ${orderClause} LIMIT ${pageSize} OFFSET ${offset}`;
  const items = asRows<TransactionRow>(queryAll(sql, params)).map((r): TransactionListItem => ({
    ...r,
    categoryId: r.category_id,
    accountId: r.account_id,
    toAccountId: r.transfer_to_account_id || null,
    currency: normalizeCurrency(r.currency),
    originalAmount: Number(r.original_amount) > 0 ? Number(r.original_amount) : Number(r.amount) || 0,
    fxRate: String(r.fx_rate || '1'),
    fxFee: Number(r.fx_fee) || 0,
    twdAmount: Number(r.twd_amount) || Number(r.amount) || 0,
    excludeFromStats: r.exclude_from_stats === 1,
    isFxFee: r.is_fx_fee === 1,
    aiCreated: r.ai_created === 1,
    noteAiModified: r.note_ai_modified === 1,
    repaymentSummaryId: (r as { repayment_summary_id?: string }).repayment_summary_id || '',
    linkedId: r.linked_id || '',
    sourceRecurringId: r.source_recurring_id || null,
    sourceRecurringName: r.source_recurring_id ? (r.source_recurring_name || null) : null,
    scheduledDate: r.scheduled_date || null,
    createdAt: r.created_at,
    updatedAt: r.updated_at,
    attachmentCount: Number(r.attachment_count) || 0,
    firstAttachmentId: typeof r.first_attachment_id === 'string' ? r.first_attachment_id : null,
  }));

  const response: TransactionListResponse = {
    data: items,
    items,
    total,
    page: pageNum,
    pageSize,
    totalPages: Math.ceil(total / pageSize),
    sort: `${sortField}_${sortDir.toLowerCase()}`,
  };

  return NextResponse.json(response);
}

async function handlePOST(request: NextRequest) {
  const auth = await requireAuth(request);
  if (auth instanceof NextResponse) return auth;

  const body = await request.json().catch(() => ({})) as CreateTransactionRequest;
  const { type, amount, categoryId, accountId, note, excludeFromStats } = body;
  const rawDate = body.date;

  const date = (rawDate == null || String(rawDate).trim() === '')
    ? todayInUserTz(auth.userTimezone)
    : normalizeDate(rawDate);
  if (!date) return NextResponse.json({ error: '日期格式無效' }, { status: 400 });
  if (!isValidIsoDate(date)) return NextResponse.json({ error: '日期格式無效', code: 'ValidationError', field: 'date' }, { status: 400 });
  if (!type || !TRANSACTION_TYPES.has(type)) {
    return NextResponse.json({ error: '交易類型無效' }, { status: 400 });
  }

  if (categoryId) {
    const catOwned = queryOne('SELECT id FROM categories WHERE id = ? AND user_id = ?', [categoryId, auth.userId]);
    if (!catOwned) return NextResponse.json({ error: '分類不存在或無權限' }, { status: 400 });
    const catRow = asRow<{ parent_id: string | null }>(queryOne('SELECT parent_id FROM categories WHERE id = ? AND user_id = ?', [categoryId, auth.userId]));
    if (catRow && !catRow.parent_id) {
      return NextResponse.json({ error: '交易必須指派至子分類，不能直接掛在父分類底下' }, { status: 400 });
    }
  }
  if (accountId) {
    const accOwned = asRow<{ id: string; category: string | null; account_type: string | null; is_active: number | null }>(queryOne('SELECT id, category, account_type, is_active FROM accounts WHERE id = ? AND user_id = ?', [accountId, auth.userId]));
    if (!accOwned) return NextResponse.json({ error: '帳戶不存在或無權限' }, { status: 400 });
    const isCreditCard = accOwned.category === 'credit_card' || accOwned.account_type === '信用卡';
    if (type === 'expense' && isCreditCard && accOwned.is_active === 0) {
      return NextResponse.json({ error: '此信用卡已停用，無法新增刷卡消費', code: 'CreditCardDisabled' }, { status: 409 });
    }
  }

  const numAmt = Number(body.originalAmount ?? amount);
  if (!Number.isFinite(numAmt) || numAmt <= 0) {
    return NextResponse.json({ error: '金額必須大於 0', code: 'ValidationError', field: 'amount' }, { status: 400 });
  }

  let converted;
  try {
    converted = convertToTwd(Number(body.originalAmount ?? amount), body.currency || 'TWD', body.fxRate, auth.userId);
  } catch (e) {
    return NextResponse.json({ error: e instanceof Error ? e.message : '金額格式錯誤' }, { status: 400 });
  }

  // 國外刷卡手續費另存為獨立交易，故原交易 twd_amount 不含手續費（fx_fee=0）。
  const fxFee = resolveOverseasFee({
    userId: auth.userId,
    accountId: accountId || null,
    currency: converted.currency,
    twdBase: converted.twdAmount,
    clientFxFee: body.fxFee,
  });
  const twdAmountInt = computeTwdAmount(
    Math.round(converted.originalAmount * 100) / 100,
    converted.fxRate,
    0
  );

  const result = insertIncomeExpenseTransaction({
    userId: auth.userId,
    type,
    twdAmount: twdAmountInt,
    currency: converted.currency,
    originalAmount: converted.originalAmount,
    fxRate: converted.fxRate,
    fxFee,
    date,
    categoryId: categoryId || null,
    accountId: accountId || null,
    note: note || '',
    excludeFromStats: !!excludeFromStats,
    clientRef: typeof body.clientRef === 'string' && CLIENT_REF_REGEX.test(body.clientRef) ? body.clientRef : undefined,
  });

  // Webhook：交易新增事件（issue #258）。僅在本次真的新增列時發出，
  // 離線重送（client_ref 去重，inserted=false）不重複發出，維持事件冪等。
  // emit 為盡力而為，失敗不影響交易本身。
  if (result.inserted) {
    emitTransactionEvent(auth.userId, 'transaction.created', {
      id: result.id,
      type,
      amount: twdAmountInt,
      currency: converted.currency,
      date,
      account_id: accountId || null,
      category_id: categoryId || null,
      note: note || '',
    });
  }

  return NextResponse.json(result, { status: 201 });
}

export const POST = withLedgerWriteAudit(handlePOST);