# asset_openapi.yaml
openapi: "3.2.0"
info:
  title: AssetPilot API
  version: "4.110.1"
  description: |
    個人資產管理平台 AssetPilot 的完整 API Schema。
    供 Cloudflare API Shield Schema Validation 使用。
    認證方式：JWT 儲存於 httpOnly Cookie（token），除 /api/auth/* 與 /api/config 外均需認證。
    記帳 API 可用 X-Ledger-Id header 或 ledgerId query 選擇帳本，未指定則使用個人帳本。
    每次請求皆驗證登入者的帳本成員資格；viewer 僅可 GET/HEAD/OPTIONS。
    個人偏好、投資、登入憑證及 MCP/API Token/LINE 整合不採用共享帳本選擇。
servers:
  - url: https://asset.shao.one
    description: Production

# ─────────────────────────────────────────
# 標籤定義（對應 Cloudflare API Shield Endpoint Labels）
# ─────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
  schemas:
...
    TransactionInput:
      type: object
      required: [type, amount, date, categoryId, accountId]
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        originalAmount:
          type: number
        currency:
          type: string
        fxRate:
          type: number
        date:
          type: string
        categoryId:
          type: string
        accountId:
          type: string
        note:
          type: string
        excludeFromStats:
          type: boolean
        fxFee:
          type: number
          description: 海外刷卡手續費（TWD）
        clientRef:
          type: string
          pattern: "^[a-f0-9]{32}$"
          description: |
            離線記帳的 idempotency key（007-pwa-offline-entry）。離線新增的交易由
            client 產生此值；伺服器以 (user_id, client_ref) 唯一索引去重，同一
            clientRef 重送不會產生重複交易，並回傳與新建相同的結果。線上直接新增時省略；
            格式不符時視為未提供。

    # ── 預算 ──
...
security:
  - cookieAuth: []

# ─────────────────────────────────────────
# 路由定義
# ─────────────────────────────────────────
paths:
...
  /api/transactions:
    get:
      operationId: getTransactions
      tags: [Transactions]
      summary: 取得交易記錄（支援篩選分頁）
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: string
        - name: accountId
          in: query
          schema:
            type: string
        - name: keyword
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: |
            交易清單（含分頁資訊）。每筆交易含唯讀衍生欄位 repaymentSummaryId
            （非還款交易為空字串；非空代表該筆屬於某次信用卡還款摘要，可作為開啟摘要的入口）。
    post:
      operationId: createTransaction
      tags: [Transactions]
      summary: 新增交易
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "200":
          description: 新增成功
...
  /api/transactions/batch-update:
    post:
      operationId: batchUpdateTransactions
      tags: [Transactions]
      summary: 批次更新交易（分類 / 帳戶 / 日期）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, fields]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                fields:
                  type: object
                  properties:
                    categoryId:
                      type: string
                    accountId:
                      type: string
                    date:
                      type: string
      responses:
        "200":
          description: 更新成功
...